RayTracer: Added Material, Scene, SceneObject, Sphere, and Tracer

+++ SoftwareRenderer/Source/MathUtils.h
@@ -16,7 +16,7 @@ inline void QuadraticPoint(int x1, int y1, int x2, int y2, int x3, int y3, float
 {
 	float one_minus_t = 1 - t;
 
-	float a = std::pow(one_minus_t, 2);
+	float a = (float)(std::pow(one_minus_t, 2));
 	float b = 2 * one_minus_t * t;
 	float c = t * t;
 
@@ -28,10 +28,10 @@ inline void CubicPoint(int x1, int y1, int x2, int y2, int x3, int y3, int x4, i
 {
 	float one_minus_t = 1 - t;
 
-	float a = std::pow(one_minus_t, 3);
-	float b = 3 * std::pow(one_minus_t, 2) * t;
+	float a = (float)(std::pow(one_minus_t, 3));
+	float b = (float)(3 * std::pow(one_minus_t, 2) * t);
 	float c = 3 * one_minus_t * (t * t);
-	float d = std::pow(t, 3);
+	float d = (float)(std::pow(t, 3));
 
 	x = (int)(a * x1 + b * x2 + c * x3 + d * x4);
 	y = (int)(a * y1 + b * y2 + c * y3 + d * y4);
